feat: 프로필 작성을 위한 Base 뷰컨 구현
#119
